Regenerate site with 07da38c9502f5255fa59ebe240d10f5372e92e17
SiteGenerator: Handle "github-actions" user differently

This patch updates the site generator to handle the user called
"github-actions" different from all other users. The display name on the
site will be "GitHub Actions" and clicking on the link will redirect
to the actions overview of the github repository the binary got released
in.

[skip ci]